vendor update and renaming
