+++ builtin/logical/azureservicebus/path_config_lease.go
@@ -1,4 +1,4 @@
-package azuresas
+package azureservicebus
 
 import (
 	"fmt"
