+++ builtin/logical/azureservicebus/path_roles.go
@@ -1,4 +1,4 @@
-package azuresas
+package azureservicebus
 
 import (
 	"fmt"
