+++ builtin/logical/azureservicebus/util.go
@@ -1,4 +1,4 @@
-package azuresas
+package azureservicebus
 
 import (
 	"crypto/hmac"
